refactor: ManyToMany 의 패키지 이름을 조금 더 직관적으로 변경합니다
