+++ src/main/java/nextstep/subway/applicaion/SubwayLineService.java
@@ -3,8 +3,8 @@ package nextstep.subway.applicaion;
 import nextstep.subway.applicaion.dto.subwayLine.CreateSubwayLineRequest;
 import nextstep.subway.applicaion.dto.subwayLine.SubwayLineResponse;
 import nextstep.subway.applicaion.dto.subwayLine.UpdateSubwayLineRequest;
-import nextstep.subway.domain.manyToMany.StationToSubwayLine;
-import nextstep.subway.domain.manyToMany.StationToSubwayLineRepository;
+import nextstep.subway.domain.stationToSubwayBridge.StationToSubwayLine;
+import nextstep.subway.domain.stationToSubwayBridge.StationToSubwayLineRepository;
 import nextstep.subway.domain.station.Station;
 import nextstep.subway.domain.station.StationRepository;
 import nextstep.subway.domain.subwayLine.SubwayLine;
