+++ src/main/java/nextstep/subway/domain/stationToSubwayBridge/StationToSubwayLineRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain.manyToMany;
+package nextstep.subway.domain.stationToSubwayBridge;
 
 import nextstep.subway.domain.station.Station;
 import nextstep.subway.domain.subwayLine.SubwayLine;
